requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Werkzeug==2.2.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.28.2
python-slugify==8.0.4
